Add Losseable, base funtional of the compsite method, change Buyable

+++ src/main/java/workwithzoo/sourceofprofit/Profitable.java
@@ -6,5 +6,5 @@ package workwithzoo.sourceofprofit;
  * @author DantalioNxxi
  */
 public interface Profitable {
-    float getProfit();
+    double getProfit();
 }
